requirements.txt
absl-py==0.9.0
ansicon==1.89.0
asgiref==3.2.7
astunparse==1.6.3
attrs==19.3.0
audioread==2.1.8
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.6
boto==2.49.0
boto3==1.13.23
botocore==1.17.4
cached-property==1.5.1
cachetools==4.1.0
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.42.0
django-braces==1.14.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.9.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
fabric==2.5.0
future==0.16.0
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.7
importlib-metadata==1.6.0
invoke==1.4.1
jinxed==1.0.0
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
librosa==0.7.2
llvmlite==0.32.1
Markdown==3.2.2
musicnn==0.1.0
numba==0.49.1
numpy==1.16.6
oauthlib==3.1.0
opt-einsum==3.2.1
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
prettytable==0.7.2
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pypi==2.1
pyrsistent==0.16.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
semantic-version==2.5.0
sentry-sdk==0.14.4
six==1.11.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
SoundFile==0.10.3.post1
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==2.1.0
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
